% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_controls}
\alias{extract_controls}
\title{Extract control variables from a model object}
\usage{
extract_controls(m, treatment)
}
\arguments{
\item{m}{Model object (lm, fixest, etc.)}

\item{treatment}{Character. Name of the treatment variable to exclude}
}
\value{
Character vector of control variable names
}
\description{
This function standardizes the extraction of control variables from different
model types (lm, fixest) to ensure consistency across the package.
}
\keyword{internal}
