% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extreme_coef_after_k_flips}
\alias{extreme_coef_after_k_flips}
\title{Find extreme coefficient after k flips}
\usage{
extreme_coef_after_k_flips(
  cont,
  bin,
  k,
  treatment,
  dat,
  setting,
  direction = "max",
  m
)
}
\arguments{
\item{cont}{Numeric vector. Continuous variable (residuals from model)}

\item{bin}{Numeric vector. Binary variable (treatment from model)}

\item{k}{Integer. Number of observations to flip}

\item{treatment}{Character. Name of treatment variable in model}

\item{dat}{Data frame. The dataset}

\item{setting}{Character. Whether variable is treatment ('d') or dependent ('y')}

\item{direction}{Character. Direction to optimize ('max' or 'min')}

\item{m}{Model object. The fitted model}
}
\value{
List containing the extreme coefficient and flip information
}
\description{
This function identifies which observations will return the most extreme positive
or negative relationship between a continuous variable (cont) and a binary variable (bin)
by strategically flipping k observations to maximize or minimize the treatment coefficient.
}
\keyword{internal}
