% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futureproofR-package.R
\docType{package}
\name{futureproofR}
\alias{futureproofR-package}
\alias{futureproofR}
\title{Sensitivity analysis on binary regression coefficients}
\description{
Sensitivity analysis on regression coefficients to assess how measurement error
(misclassification) in a binary variable affects the estimated treatment effect.
It calculates extreme bounds on the coefficient by flipping binary classifications
in the data.
}
\author{
Leonardo Dantas \email{leodantas@princeton.edu}
}
\keyword{package}
