% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignettes.R
\docType{vignette}
\name{misclassification-vignette}
\alias{misclassification-vignette}
\title{Visualization of Misclassification Impact on Treatment Effect Bias}
\description{
This vignette demonstrates how to use `futureproofR`'s `contour_plot_data()` and
`contour_plot()` functions to visualize the impact of misclassification rates and
treatment proportions on the bias of treatment effect estimates.
}
