% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prog_bar_lapply}
\alias{prog_bar_lapply}
\title{Progress bar version of lapply}
\usage{
prog_bar_lapply(X, FUN, ...)
}
\arguments{
\item{X}{Vector or list to apply function over}

\item{FUN}{Function to apply to each element of X}

\item{...}{Additional arguments passed to FUN}
}
\value{
List of results from applying FUN to each element of X
}
\description{
This function creates a self-updating progress bar akin to that of `pblapply` to
indicate percentage and time passed of long iterations. Results are returned
as a list.
}
\keyword{internal}
