% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignettes.R
\docType{vignette}
\name{sensitivity-analysis-vignette}
\alias{sensitivity-analysis-vignette}
\title{Sensitivity Analysis for Reclassified Binary Variables}
\description{
This vignette demonstrates how to use `futureproofR`'s `misclass_sens()` to perform
sensitivity analysis on binary regression coefficients, accounting for measurement error
(misclassification).
}
